requirements.txt
asyncpg
fastapi>=0.115.0
uvicorn==0.25.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pydantic>=2.6.4
python-multipart>=0.0.9
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4
tzdata>=2024.2
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
requests>=2.31.0
sqlalchemy>=2.0
psycopg2-binary
pandas>=2.2.0
numpy>=1.26.0
jq>=1.6.0
typer>=0.9.0
asyncpg==0.31.0
httpx==0.28.1
psycopg2-binary==2.9.11
SQLAlchemy==2.0.48
httpx
google-genai>=0.3.0
tenacity
greenlet
# ─── Pipeline dependencies ────────────────────────────────────────────────────
celery>=5.3.0
redis>=5.0.0
rapidfuzz>=3.6.0
serpapi>=0.1.5
openpyxl>=3.1.0
# ─── OpenAI Agents ───────────────────────────────────────────────────────────
openai>=2.0.0
openai-agents>=0.13.0
pyhunter>=2.1.0